% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_flower.R
\name{flowerp}
\alias{flowerp}
\title{A map of Gotland with Orchid genus and rain fall}
\usage{
flowerp(x1)
}
\arguments{
\item{x1}{insert year}
}
\description{
A map of Gotland with Orchid genus and rain fall
}
\details{
This function makes a map of orchid genus's in Gotland and makes a denisty plot of rain fall for that same year.
}
\section{Warning}{

Might work, might not. We hope so
}

