% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{orchidsmap}
\alias{orchidsmap}
\title{A map of Gotland with Orchid observations}
\usage{
orchidsmap(my_df, x1, y1, z1)
}
\arguments{
\item{my_df}{A data frame}

\item{x1}{Longitude}

\item{y1}{Latitude}

\item{z1}{Observation to color by, example Species, Genus, Temperature etc.}
}
\description{
A map of Gotland with Orchid observations
}
\details{
This function makes a map of Gotland and adds latitude and longitude data points of some observation.
}
\section{Warning}{

Might work, might not. We hope so
}

